refactor: newauth 패키지 네이밍 auth 로 변경
