+++ src/main/java/com/commerce/backendserver/auth/application/dto/AuthTokenResponse.java
@@ -1,0 +1,7 @@
+package com.commerce.backendserver.auth.application.dto;
+
+public record AuthTokenResponse(
+	String accessToken,
+	String refreshToken
+) {
+}
