+++ src/main/java/com/commerce/backendserver/auth/domain/OAuthUriGenerator.java
@@ -1,0 +1,10 @@
+package com.commerce.backendserver.auth.domain;
+
+import com.commerce.backendserver.auth.domain.model.OAuthProvider;
+
+public interface OAuthUriGenerator {
+
+	boolean isSupported(final OAuthProvider provider);
+
+	String generate();
+}
