+++ src/main/java/com/commerce/backendserver/auth/domain/model/AuthToken.java
@@ -1,0 +1,7 @@
+package com.commerce.backendserver.auth.domain.model;
+
+public record AuthToken(
+	String accessToken,
+	String refreshToken
+) {
+}
