+++ src/main/java/com/commerce/backendserver/auth/domain/model/OAuthMemberInfo.java
@@ -1,0 +1,10 @@
+package com.commerce.backendserver.auth.domain.model;
+
+public interface OAuthMemberInfo {
+
+	String oauthId();
+
+	String nickname();
+
+	String email();
+}
