+++ src/main/java/com/commerce/backendserver/auth/domain/model/OAuthTokenInfo.java
@@ -1,0 +1,6 @@
+package com.commerce.backendserver.auth.domain.model;
+
+public interface OAuthTokenInfo {
+
+	String accessToken();
+}
