+++ src/main/java/com/commerce/backendserver/auth/presentation/dto/OAuthLoginRequest.java
@@ -1,0 +1,7 @@
+package com.commerce.backendserver.auth.presentation.dto;
+
+public record OAuthLoginRequest(
+	String code,
+	String state
+) {
+}
