--- src/main/java/com/commerce/backendserver/newauth/application/dto/AuthTokenResponse.java
@@ -1,7 +1,0 @@
-package com.commerce.backendserver.newauth.application.dto;
-
-public record AuthTokenResponse(
-	String accessToken,
-	String refreshToken
-) {
-}
