--- src/main/java/com/commerce/backendserver/newauth/domain/AuthTokenManager.java
@@ -1,12 +1,0 @@
-package com.commerce.backendserver.newauth.domain;
-
-import com.commerce.backendserver.newauth.domain.model.AuthToken;
-
-public interface AuthTokenManager {
-
-	AuthToken generate(final Long memberId);
-
-	Long getId(final String token);
-
-	void validateToken(final String token);
-}
