--- src/main/java/com/commerce/backendserver/newauth/domain/OAuthUriGenerator.java
@@ -1,10 +1,0 @@
-package com.commerce.backendserver.newauth.domain;
-
-import com.commerce.backendserver.newauth.domain.model.OAuthProvider;
-
-public interface OAuthUriGenerator {
-
-	boolean isSupported(final OAuthProvider provider);
-
-	String generate();
-}
