--- src/main/java/com/commerce/backendserver/newauth/domain/model/AuthToken.java
@@ -1,7 +1,0 @@
-package com.commerce.backendserver.newauth.domain.model;
-
-public record AuthToken(
-	String accessToken,
-	String refreshToken
-) {
-}
