--- src/main/java/com/commerce/backendserver/newauth/domain/model/OAuthMemberInfo.java
@@ -1,10 +1,0 @@
-package com.commerce.backendserver.newauth.domain.model;
-
-public interface OAuthMemberInfo {
-
-	String oauthId();
-
-	String nickname();
-
-	String email();
-}
