--- src/main/java/com/commerce/backendserver/newauth/domain/model/OAuthTokenInfo.java
@@ -1,6 +1,0 @@
-package com.commerce.backendserver.newauth.domain.model;
-
-public interface OAuthTokenInfo {
-
-	String accessToken();
-}
