--- src/main/java/com/commerce/backendserver/newauth/presentation/dto/OAuthLoginRequest.java
@@ -1,7 +1,0 @@
-package com.commerce.backendserver.newauth.presentation.dto;
-
-public record OAuthLoginRequest(
-	String code,
-	String state
-) {
-}
